requirements.txt
# Core Dependencies
python-dateutil>=2.8.2
pyyaml>=6.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Data sources
yfinance>=0.2.28
finvizfinance>=0.14.0
alpha-vantage>=2.3.1

# Technical indicators
pandas-ta>=0.3.14b
ta-lib>=0.4.28  # Note: Requires binary installation on Windows


# Backtesting
backtrader>=1.9.78.123

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0

# Dashboard
streamlit>=1.28.0

# Alerts
python-telegram-bot>=20.0
secure-smtplib>=0.1.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

